organize markdown
