+++ docs/src/routes/glyphs.tsx
@@ -1,4 +1,4 @@
-import {Markdown} from '~/components/markdown';
+import {Markdown} from '~/components/markdown/markdown';
 
 function glyphs() {
     const md = `
