+++ docs/src/routes/types.tsx
@@ -1,4 +1,4 @@
-import {Markdown} from '~/components/markdown';
+import {Markdown} from '~/components/markdown/markdown';
 
 function types() {
     const md = `
